create custom hook that fetches user data on first mount and user changes, and implement asynchronous functions using object de construction and asyn await functions
